refactor: invoke same url when updating general and appearance settings
