+++ client/app/constants/author-appearance.js
@@ -1,0 +1,20 @@
+const optionTypes = {
+    coverStyle: "coverStyle",
+    blogLayoutStyle: "blogLayoutStyle",
+};
+
+const validOptions = {
+    [optionTypes.coverStyle]: [
+        "full",
+        "condensed",
+    ],
+    [optionTypes.blogLayoutStyle]: [
+        "vertical",
+        "cards",
+    ],
+};
+
+export {
+    validOptions,
+    optionTypes,
+};
