+++ client/app/constants/index.js
@@ -1,0 +1,6 @@
+import { validOptions, optionTypes } from "./author-appearance";
+
+export {
+    validOptions as authorAppearanceValidOptions,
+    optionTypes as authorAppearanceOptionTypes,
+};
