+++ config/routes.rb
@@ -14,7 +14,6 @@ Rails.application.routes.draw do
       post 'domain_request'
       post 'delete_domain'
       post 'delete_all_data'
-      put 'appearance' => 'authors#update_appearance'
     end
 
     resources :guestbook_entries, path: :guestbook do
